component names refactor
