+++ src/components/HeroSection.vue
@@ -1,0 +1,17 @@
+<template>
+  <section class="hero">
+    <h1 class="hero__title">Welcome to the best Cinema in Monterail</h1>
+  </section>
+</template>
+
+<script>
+export default {};
+</script>
+
+<style lang="scss" scoped>
+.hero {
+  &__title {
+    font-size: 20px;
+  }
+}
+</style>
